menteeAll 폴더명 변경, svg를 img로 변경
