+++ components/menteeall/Filter.module.scss
@@ -1,0 +1,46 @@
+.container {
+  position: absolute;
+  top: 278.5px;
+  width: 100%;
+  padding: 0 20px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  span {
+    font-family: Noto Sans KR;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 14px;
+    line-height: 22px;
+    color: #2c343f;
+  }
+  .selectBox {
+    width: 68px;
+    height: 25px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border: 1px solid #e8eaef;
+    box-sizing: border-box;
+    border-radius: 8px;
+    background-color: #fff;
+    select {
+      -o-appearance: none;
+      -webkit-appearance: none;
+      -moz-appearance: none;
+      appearance: none;
+      background: url("/images/menteeall/filterarrow.svg") no-repeat;
+      font-family: Noto Sans KR;
+      font-style: normal;
+      font-weight: 500;
+      font-size: 12px;
+      line-height: 17px;
+      color: #434a58;
+      border: none;
+      padding-left: 14px;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+    }
+  }
+}
