+++ components/menteeall/Lecture.module.scss
@@ -21,7 +21,7 @@
           rgba(0, 0, 0, 0.45),
           rgba(0, 0, 0, 0.45)
         ),
-        url("../../public/images/tuteeall/lectureimg.png");
+        url("../../public/images/menteeall/lectureimg.png");
       border-radius: 8px 8px 0px 0px;
       .category_box {
         display: flex;
