+++ components/menteeall/NavBar.module.scss
@@ -1,8 +1,7 @@
 .container {
   position: fixed;
-  left: 0;
   bottom: 0;
-  width: 100%;
+  width: 375px;
   height: 96px;
   background: #ffffff;
   box-sizing: border-box;
