+++ components/menteeall/TopMenu.module.scss
@@ -25,9 +25,7 @@
       -webkit-appearance: none;
       -moz-appearance: none;
       appearance: none;
-      background: url("data:image/svg+xml,%3Csvg width='8' height='6' viewBox='0 0 8 6' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M1 1.5L3.63099 4.33229C3.85155 4.56972 4.2142 4.55282 4.41479 4.29575L7 1.5H1Z' fill='%232C343F' stroke='%232C343F' stroke-width='1.5' stroke-linecap='round' stroke-linejoin='round'/%3E%3C/svg%3E")
-        no-repeat 94px;
-
+      background: url("/images/menteeall/location_arrow.svg") no-repeat 94px;
       color: #2c343f;
       padding-right: 16px;
     }
