--- components/tuteeall/Filter.module.scss
@@ -1,50 +1,0 @@
-
-.container {
-    position: absolute;
-    top: 278.5px;
-    width: 100%;
-    padding: 0 20px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    span {
-        font-family: Noto Sans KR;
-        font-style: normal;
-        font-weight: 500;
-        font-size: 14px;
-        line-height: 22px;
-        color: #2C343F;
-    }
-    .selectBox {
-        width: 68px;
-        height: 25px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        border: 1px solid #E8EAEF;
-        box-sizing: border-box;
-        border-radius: 8px;
-        background-color: #fff;
-        select {
-            -o-appearance: none;
-            -webkit-appearance: none;
-            -moz-appearance: none;
-            appearance: none;
-            background: url("data:image/svg+xml,%3Csvg width='10' height='17' viewBox='0 0 10 17' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cg clip-path='url(%23clip0_1148_2381)'%3E%3Cpath d='M3 13.5002L2.99976 14.0002L1.99976 13.9998L2 13.4998L3 13.5002ZM2.00366 5.96028L2.0039 5.46028L3.0039 5.46077L3.00366 5.96077L2.00366 5.96028ZM1.34998 7.33078L0.992882 7.68075L0.292929 6.96656L0.650024 6.61659L1.34998 7.33078ZM2.50366 5.5L2.15368 5.14291L2.50451 4.79907L2.8545 5.14376L2.50366 5.5ZM4.35085 6.61744L4.70709 6.96829L4.0054 7.68077L3.64915 7.32992L4.35085 6.61744ZM2 13.4998L2.00366 5.96028L3.00366 5.96077L3 13.5002L2 13.4998ZM0.650024 6.61659L2.15368 5.14291L2.85363 5.85709L1.34998 7.33078L0.650024 6.61659ZM2.8545 5.14376L4.35085 6.61744L3.64915 7.32992L2.15281 5.85624L2.8545 5.14376Z' fill='%23434A58'/%3E%3Cpath d='M7.09961 5.49976L7.09985 4.99976L8.09985 5.00024L8.09961 5.50024L7.09961 5.49976ZM8.09595 13.0397L8.09571 13.5397L7.09571 13.5392L7.09595 13.0392L8.09595 13.0397ZM8.74963 11.6692L9.10673 11.3192L9.80668 12.0334L9.44959 12.3834L8.74963 11.6692ZM7.59595 13.5L7.94593 13.8571L7.5951 14.2009L7.24511 13.8562L7.59595 13.5ZM5.74876 12.3826L5.39252 12.0317L6.09421 11.3192L6.45045 11.6701L5.74876 12.3826ZM8.09961 5.50024L8.09595 13.0397L7.09595 13.0392L7.09961 5.49976L8.09961 5.50024ZM9.44959 12.3834L7.94593 13.8571L7.24598 13.1429L8.74963 11.6692L9.44959 12.3834ZM7.24511 13.8562L5.74876 12.3826L6.45045 11.6701L7.9468 13.1438L7.24511 13.8562Z' fill='%23434A58'/%3E%3C/g%3E%3Cdefs%3E%3CclipPath id='clip0_1148_2381'%3E%3Crect width='10' height='17' fill='white'/%3E%3C/clipPath%3E%3C/defs%3E%3C/svg%3E%0A") no-repeat ;
-            font-family: Noto Sans KR;
-            font-style: normal;
-            font-weight: 500;
-            font-size: 12px;
-            line-height: 17px;
-            color: #434A58;
-            border: none;
-            padding-left: 14px;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            
-           
-        }
-    }
-    
-}
