--- components/tuteeall/ManImage.jsx
@@ -1,239 +1,0 @@
-import React from "react";
-import styles from "./ManImage.module.scss";
-
-function ManImage() {
-  return (
-    <svg
-      className={styles.container}
-      width="133"
-      height="92"
-      viewBox="0 0 133 92"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M45.4988 51.5817L49.3994 85H47.4491L43.5484 51L45.4988 51.5817Z"
-        fill="url(#paint0_linear_490_11809)"
-      />
-      <path
-        d="M0.97541 50L43.5486 51L47.4492 85L3.7868 83L0.0060997 51.0762C-0.0582619 50.5022 0.394215 50 0.97541 50Z"
-        fill="url(#paint1_linear_490_11809)"
-      />
-      <path
-        d="M67.6857 36.1664C67.9368 35.7864 67.978 35.3057 67.7902 34.891C65.4781 29.7825 65.4096 25.6222 66.0363 22.5864C66.0662 22.5326 66.0969 22.4771 66.1294 22.4172C66.1294 22.4172 68.0685 15.5483 82.2054 18.2327C83.5327 18.4852 84.6642 19.3285 85.2918 20.5128C85.9756 21.8038 87.4389 23.5521 90.5612 24.9004C91.396 25.2604 92.1158 25.8391 92.6601 26.561L93.0788 27.1171C93.0788 27.1171 94.7466 29.4458 95.1416 37.1243C94.6193 43.8318 87.8032 53.5206 83.0674 53.354C83.0253 53.3523 82.9823 53.3523 82.9402 53.3488C78.2 53.0017 74.3377 46.7549 73.6355 44.8461C72.9332 42.9373 69.0709 44.2388 66.5253 42.677C64.4949 41.4303 66.7061 37.6483 67.6857 36.1664Z"
-        fill="url(#paint2_linear_490_11809)"
-      />
-      <path
-        d="M83.0659 53.3526C87.8016 53.5192 94.6186 43.8305 95.14 37.1229C94.745 29.4444 93.0772 27.1157 93.0772 27.1157L93.6794 25.5375C93.9541 24.8174 94.3272 24.1389 94.7907 23.5246C95.004 23.2427 95.234 22.9728 95.4859 22.7212C96.857 21.3504 99.0217 20.1947 101.153 23.1975C101.153 23.1975 103.611 26.0607 99.7836 34.4723C99.73 34.5903 99.6677 34.7031 99.608 34.8185C99.4448 35.1317 99.0893 36.4027 101.444 39.5678C101.484 39.6208 101.532 39.6659 101.562 39.7257C104.839 46.363 108.701 52.2629 108.701 52.2629L87.1951 75.0813L83.0641 53.3543L83.0659 53.3526Z"
-        fill="url(#paint3_linear_490_11809)"
-      />
-      <path
-        d="M109.624 13.9572C112.398 27.9086 102.811 38.1864 101.445 39.5677C99.0895 36.4034 99.445 35.1315 99.6083 34.8183C99.668 34.7038 99.7303 34.5901 99.7838 34.4721C103.611 26.0605 101.153 23.1974 101.153 23.1974C100.773 22.6629 100.393 22.2604 100.016 21.9671C103.943 15.7419 106.166 10.5535 107.139 8.02441C109.04 11.1322 109.624 13.9572 109.624 13.9572Z"
-        fill="url(#paint4_linear_490_11809)"
-      />
-      <path
-        d="M62.1786 15.3265C64.4714 14.6767 66.9626 12.8295 66.9626 12.8295L89.6099 0.249004C99.3104 -1.16348 104.496 3.70387 107.139 8.02463C106.166 10.5537 103.943 15.7421 100.015 21.9673C98.2755 20.6138 96.613 21.5942 95.4859 22.7213C95.2339 22.9729 95.004 23.2427 94.7906 23.5247C94.3272 24.139 93.9541 24.8174 93.6793 25.5376L93.0772 27.1158L92.6585 26.5596C92.1142 25.8369 91.3944 25.2591 90.5596 24.899C90.4526 24.853 90.3472 24.8053 90.2436 24.7584C89.8811 24.5927 89.5414 24.4218 89.2245 24.2466C87.0028 23.0197 85.8687 21.6029 85.2902 20.5114C84.6626 19.328 83.532 18.4838 82.2039 18.2313C68.0669 15.5469 66.1278 22.4159 66.1278 22.4159C66.0953 22.4749 66.0646 22.5313 66.0348 22.5851C66.0444 22.5382 66.0541 22.4914 66.0637 22.4445C63.4303 21.2246 61.9126 19.5501 61.0462 18.1437C60.387 17.0722 60.9576 15.6727 62.1786 15.3265Z"
-        fill="url(#paint5_linear_490_11809)"
-      />
-      <path
-        d="M67.8463 28.4528C68.136 28.3157 68.4388 28.2325 68.7417 28.1882C69.7029 28.0477 70.6667 28.3001 71.2337 28.4988C72.0369 29.1582 72.8428 30.1291 73.5046 31.3082C73.7188 31.6899 73.9058 32.0726 74.0664 32.4508L70.2559 30.8857L67.6356 29.8098C67.6216 29.2831 67.6909 28.8207 67.8463 28.4528Z"
-        fill="#F6F8F8"
-      />
-      <path
-        d="M74.3395 30.5635L89.2253 24.2461C89.5422 24.4214 89.8819 24.5923 90.2444 24.758L74.7523 31.3328C75.9684 33.8158 76.0451 36.2161 74.8662 36.8627L74.0112 37.3772C74.8153 36.8523 74.9715 35.4406 74.523 33.7644L67.7841 30.975C67.6954 30.5629 67.6454 30.1707 67.6357 29.8098L74.0665 32.4508C73.9059 32.0725 73.7198 31.6899 73.5047 31.3081C72.8429 30.129 72.037 29.1582 71.2338 28.4988C70.6668 28.3001 69.703 28.0476 68.7418 28.1882C68.4389 28.2324 68.1361 28.3157 67.8464 28.4528C67.986 28.1222 68.1958 27.8689 68.4767 27.7153L69.3975 27.2104C69.7188 27.0342 70.1032 27.0021 70.5228 27.0941C71.7281 27.3591 73.2239 28.6569 74.3395 30.5635Z"
-        fill="url(#paint6_linear_490_11809)"
-      />
-      <path
-        d="M67.7832 30.9756C67.9755 31.8684 68.3485 32.8549 68.898 33.8344C68.9628 33.9497 69.0289 34.0631 69.0963 34.1744C68.5069 34.3469 67.618 34.6606 67.2942 35.0581C67.1423 35.2446 67.1722 35.5179 67.3609 35.668C67.4408 35.7322 67.5382 35.7635 67.6348 35.7635C67.7629 35.7635 67.8902 35.7079 67.9771 35.6021C68.1286 35.4159 68.9079 35.1196 69.599 34.9355C71.0392 36.9364 72.8397 38.023 73.9261 37.4272L74.0095 37.3769C74.8135 36.852 74.9698 35.4404 74.5212 33.7641L67.7832 30.9756Z"
-        fill="url(#paint7_linear_490_11809)"
-      />
-      <path
-        d="M79.7361 45.8423C79.8055 45.8423 79.8757 45.8258 79.9415 45.7919C80.003 45.7598 81.4654 44.9851 83.0086 43.4598C85.0635 41.4287 86.2687 39.0254 86.4934 36.5101C86.5145 36.2715 86.3363 36.0607 86.0949 36.0399C85.8544 36.0182 85.6402 36.1952 85.6191 36.4338C85.1205 42.0221 79.586 44.9963 79.5298 45.025C79.3156 45.1378 79.2349 45.3998 79.3481 45.6115C79.4271 45.759 79.579 45.8423 79.7361 45.8423Z"
-        fill="url(#paint8_linear_490_11809)"
-      />
-      <path
-        d="M126.166 61.9287C129.55 65.0747 131 67.9665 131 67.9665C136.881 77.8574 127.927 91.3923 127.927 91.3923H116.859C126.653 79.0738 126.67 66.8689 126.166 61.9287Z"
-        fill="url(#paint9_linear_490_11809)"
-      />
-      <path
-        d="M78.2424 67.2725C79.3264 66.6183 82.0617 66.9428 83.6733 67.2117L83.9542 81.4147L104.806 61.98L95.8835 67.2725L93.5126 65.178L108.086 51.2987L106.477 48.6768L114.583 55.3861C119.757 56.9009 123.542 59.4917 126.164 61.9288C126.669 66.8682 126.651 79.073 116.857 91.3923H56.9996C56.9996 91.3923 76.6623 68.2269 78.2424 67.2725Z"
-        fill="url(#paint10_linear_490_11809)"
-      />
-      <path
-        d="M78.358 52.3887L75.8396 54.0797L79.7353 54.779L80.9062 53.3543L78.358 52.3887Z"
-        fill="#C8908C"
-      />
-      <path
-        d="M104.807 61.9805L83.9545 81.4152L83.6736 67.2122C84.4118 67.3354 84.9139 67.4465 84.9139 67.4465L86.5703 71.7907L93.5128 65.1785L95.8838 67.273L104.807 61.9805Z"
-        fill="url(#paint11_linear_490_11809)"
-      />
-      <path
-        d="M75.609 54.2422L73.4549 55.8195L78.0669 56.9483L79.208 55.2998L75.609 54.2422Z"
-        fill="#C8908C"
-      />
-      <path
-        d="M72.6749 56.415L70.5182 58.1269L73.8723 59.358L77.9795 59.2903V57.208L72.6749 56.415Z"
-        fill="#C8908C"
-      />
-      <path
-        d="M76.7509 57.9892C76.7483 57.9892 75.3403 57.9892 75.3377 57.9892L70.8705 57.8417C70.4018 58.2208 69.9225 58.6139 69.4327 59.0234C69.2905 59.1423 69.1474 59.2585 69.0034 59.3748C68.439 59.8329 66.0259 62.1269 69.114 65.7969C71.7141 68.8874 71.371 69.1972 72.0004 69.5C72.4814 69.7317 73.2722 69.7941 73.4056 69.2822C73.3775 69.3689 73.4284 69.1937 73.4056 69.2822C73.4056 69.2822 74.0297 65.1601 73.1519 64.3506C72.2741 63.5411 75.7002 63.7719 74.4695 70.0335C74.5125 70.1914 74.552 70.3398 74.5863 70.4769C75.3956 73.6715 75.6194 76.9823 75.2718 80.2576C74.7249 85.4086 74.2299 91.3484 74.7697 91.9991H94.2402C94.2402 91.9991 91.5226 79.4515 89.0191 71.8009C88.3423 69.7325 87.9525 67.5842 87.8674 65.4126C87.6427 59.6837 87.0853 48.913 86.0556 48.5312C86.0556 48.5312 83.1527 49.1932 77.411 53.0047L79.7934 53.8524L76.5578 53.5799C75.9574 53.9912 75.328 54.4354 74.6714 54.9143L78.6821 55.7316L73.9955 55.4124C73.2002 56.0041 72.3672 56.6444 71.4938 57.3393L76.9265 57.9866"
-        fill="url(#paint12_linear_490_11809)"
-      />
-      <path
-        d="M73.1517 64.3509C72.2739 63.5414 75.6999 63.7722 74.4693 70.0337C74.4298 69.8889 74.3859 69.7353 74.3367 69.5713C73.9294 68.2117 73.6055 68.6638 73.4054 69.2824C73.4054 69.2824 74.0295 65.1612 73.1517 64.3509Z"
-        fill="#C8908C"
-      />
-      <path
-        d="M55.5398 32.1612L55.4026 30.2192C55.2253 28.9593 55.7283 27.6995 56.7247 26.9077L58.1616 25.5942L55.9676 25.6202C54.7888 25.7213 53.6362 25.2351 52.8855 24.3207L51.5946 22.9723L51.6832 24.8369C51.8294 26.0109 51.3878 27.1815 50.5026 27.9665L48.9727 29.5393L50.9147 29.4022C52.1747 29.2249 53.4345 29.7279 54.2262 30.7243L55.5398 32.1612Z"
-        fill="#CCD5EB"
-      />
-      <defs>
-        <linearGradient
-          id="paint0_linear_490_11809"
-          x1="43.4999"
-          y1="51"
-          x2="49.9999"
-          y2="81"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#334163" />
-          <stop offset="1" stop-color="#1C2539" />
-        </linearGradient>
-        <linearGradient
-          id="paint1_linear_490_11809"
-          x1="-1.2367e-06"
-          y1="50"
-          x2="46"
-          y2="72.5"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#6F7C9B" />
-          <stop offset="1" stop-color="#384461" />
-        </linearGradient>
-        <linearGradient
-          id="paint2_linear_490_11809"
-          x1="93.5"
-          y1="43"
-          x2="66"
-          y2="23"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#F3C6C3" />
-          <stop offset="1" stop-color="#FFE1DF" />
-        </linearGradient>
-        <linearGradient
-          id="paint3_linear_490_11809"
-          x1="84.4993"
-          y1="42.5"
-          x2="107.499"
-          y2="51"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#E9ABA5" />
-          <stop offset="1" stop-color="#F8CAC7" />
-        </linearGradient>
-        <linearGradient
-          id="paint4_linear_490_11809"
-          x1="101.5"
-          y1="19"
-          x2="102"
-          y2="39.5"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#28263D" />
-          <stop offset="1" stop-color="#6C6A89" />
-        </linearGradient>
-        <linearGradient
-          id="paint5_linear_490_11809"
-          x1="102.499"
-          y1="17.5"
-          x2="60.9993"
-          y2="17"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#42405C" />
-          <stop offset="1" stop-color="#0A0720" />
-        </linearGradient>
-        <linearGradient
-          id="paint6_linear_490_11809"
-          x1="89.5"
-          y1="23.5"
-          x2="68.5"
-          y2="33.5"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#C7D5FA" />
-          <stop offset="1" stop-color="#A9BDF4" />
-        </linearGradient>
-        <linearGradient
-          id="paint7_linear_490_11809"
-          x1="67.4991"
-          y1="32"
-          x2="73.4991"
-          y2="37.5"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="white" />
-          <stop offset="1" stop-color="#E2E2E2" />
-        </linearGradient>
-        <linearGradient
-          id="paint8_linear_490_11809"
-          x1="86.5"
-          y1="36"
-          x2="79.5"
-          y2="46"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#C8908C" />
-          <stop offset="1" stop-color="#9A706C" />
-        </linearGradient>
-        <linearGradient
-          id="paint9_linear_490_11809"
-          x1="125.5"
-          y1="74"
-          x2="134"
-          y2="76.5"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#C1CFF6" />
-          <stop offset="1" stop-color="#A6BBF1" />
-        </linearGradient>
-        <linearGradient
-          id="paint10_linear_490_11809"
-          x1="119.499"
-          y1="56.5"
-          x2="64.9992"
-          y2="80.5"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#D9E3FF" />
-          <stop offset="1" stop-color="#C5D5FF" />
-        </linearGradient>
-        <linearGradient
-          id="paint11_linear_490_11809"
-          x1="102.5"
-          y1="60.5"
-          x2="88.9996"
-          y2="76"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#A8BFFC" />
-          <stop offset="1" stop-color="#AABDF0" />
-        </linearGradient>
-        <linearGradient
-          id="paint12_linear_490_11809"
-          x1="82.5001"
-          y1="50"
-          x2="81.5001"
-          y2="93.5"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#FEDAD7" />
-          <stop offset="1" stop-color="#F0B3AE" />
-        </linearGradient>
-      </defs>
-    </svg>
-  );
-}
-
-export default ManImage;
