+++ pages/menteeAll.module.scss
@@ -4,4 +4,5 @@
   position: relative;
   width: 375px;
   height: 100vh;
+  background: linear-gradient(#f0f4fe -14.72%, #ffffff 29.09%);
 }
